Split parameter_view module
