+++ crates/apps/looper/looper-vst/src/ui/bottom_panel/parameter_view/parameter_view_model.rs
